Working on EditablePage and Toolbar components
