+++ docs/changelog/changelog_en.md
@@ -5,6 +5,7 @@
 ### New
 
 - Added success and error notifications when saving changes in the editor.
+- The editor toolbar now supports all the editable pages.
 
 ### Fixed
